+++ Swift616/Calendar/Color.swift
@@ -42,7 +42,7 @@ extension UIColor {
         
         
         
-        return UIColor(red: Float(r)/255.0, green: Float(g)/255.0, blue: Float(b)/255.0, alpha: 1.0)
+        return UIColor(red: CGFloat(r)/255.0, green: CGFloat(g)/255.0, blue: CGFloat(b)/255.0, alpha: 1.0)
     }
     
 }
